// frontend/src/api/generated/models/itemAvaliacaoDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { ItemAvaliacaoDtoModoEspecial } from './itemAvaliacaoDtoModoEspecial';

export interface ItemAvaliacaoDto {
  id: string;
  materiaId: string;
  nome: string;
  /** Decimal serializado como string */
  valorMaximo: string;
  /** Decimal serializado como string */
  peso: string;
  especial: boolean;
  /** @nullable */
  modoEspecial: ItemAvaliacaoDtoModoEspecial;
  /** @nullable */
  itemSubstituidoId: string | null;
}